Crate radix_trie

A wonderful, fast, safe, generic radix trie implementation.

To get started, see the docs for Trie below.

Modules

iter

Iterators over key-value pairs, keys, values and child subtries.

Structs

NibbleVec

A data-structure for storing a sequence of 4-bit values.

SubTrie

Immutable view of a sub-tree a larger trie.

SubTrieMut

Mutable view of a sub-tree of a larger trie.

Trie

Data-structure for storing and querying string-like keys and associated values.

Traits

TrieCommon

Common functionality available for tries and subtries.

TrieKey

Trait for types which can be used to key a Radix Trie.

Type Aliases

SubTrieResult

Wrapper for subtrie lookup results.