Struct radix_trie::SubTrie

pub struct SubTrie<'a, K: 'a, V: 'a> { /* fields omitted */ }

Immutable view of a sub-tree a larger trie.

Implementations

impl<'a, K, V> SubTrie<'a, K, V> where
K: TrieKey,

pub fn get<Q: ?Sized>(&self, key: &Q) -> SubTrieResult<&V> where
K: Borrow<Q>,
Q: TrieKey,

Look up the value for the given key, which should be an extension of this subtrie’s key.

The key may be any borrowed form of the trie’s key type, but TrieKey on the borrowed form must match those for the key type