Module radix_trie::traversal

Customisable, user-implementable traversals for tries.

This module contains 4 traits that allow you to write your own recursive trie traversals. Each trait handles the key-splitting logic and result handling required to write a traversal. All of the library's important operations are implemented as traversals (insert, remove, get_ancestor, etc).

The trait is basically a massively higher-order function. The Input type allows you to define values to be passed down the traversal as it proceeds, while the Output type allows you to propogate a value back up the call stack. Look at the source for run.

Which Trait?

Traversal - immutable trie, references not allowed in Output type.

RefTraversal - immutable trie, references allowed in Output type.

TraversalMut - mutable trie, references not allowed in Output type.

RefTraversalMut - mutable trie, references allowed in Output type.

Traits

RefTraversal

RefTraversalMut

Traversal

TraversalMut