Trait radix_trie::traversal::Traversal

pub trait Traversal<'a, K: 'a, V: 'a> where K: TrieKey {
    type Key: 'a;
    type Value: 'a;
    type Result;
    fn default_result() -> Self::Result;

    fn root_fn(trie: &Trie<K, V>, key: Self::Key, value: Self::Value) -> Self::Result { ... }
    fn no_child_fn(trie: &Trie<K, V>, key: Self::Key, value: Self::Value, nv: NibbleVec, bucket: usize) -> Self::Result { ... }
    fn full_match_fn(child: &Trie<K, V>, key: Self::Key, value: Self::Value, nv: NibbleVec) -> Self::Result { ... }
    fn partial_match_fn(child: &Trie<K, V>, key: Self::Key, value: Self::Value, nv: NibbleVec, idx: usize) -> Self::Result { ... }
    fn first_prefix_fn(trie: &Trie<K, V>, key: Self::Key, value: Self::Value, nv: NibbleVec) -> Self::Result { ... }
    fn action_fn(trie: &Trie<K, V>, intermediate: Self::Result, bucket: usize) -> Self::Result { ... }
    fn run(trie: &Trie<K, V>, key: Self::Key, value: Self::Value, key_fragments: NibbleVec) -> Self::Result { ... }
}

Trait capturing a traversal of a trie.

By providing functions for each of the different cases, it is possible to describe a number of different traversals. For now it's probably best to view the source for run to understand how best to implement each function.

Associated Types

type Key: 'a

Key type to be threaded through by run, needn't be K (is often &'a K).

type Value: 'a

Value type to be threaded through by run, needn't be V (is often ()).

type Result

Type returned by the entire traversal, for insert it's Option<V>.

Required Methods

fn default_result() -> Self::Result

Provided Methods

fn root_fn(trie: &Trie<K, V>, key: Self::Key, value: Self::Value) -> Self::Result

fn no_child_fn(trie: &Trie<K, V>, key: Self::Key, value: Self::Value, nv: NibbleVec, bucket: usize) -> Self::Result

fn full_match_fn(child: &Trie<K, V>, key: Self::Key, value: Self::Value, nv: NibbleVec) -> Self::Result

fn partial_match_fn(child: &Trie<K, V>, key: Self::Key, value: Self::Value, nv: NibbleVec, idx: usize) -> Self::Result

fn first_prefix_fn(trie: &Trie<K, V>, key: Self::Key, value: Self::Value, nv: NibbleVec) -> Self::Result

fn action_fn(trie: &Trie<K, V>, intermediate: Self::Result, bucket: usize) -> Self::Result

fn run(trie: &Trie<K, V>, key: Self::Key, value: Self::Value, key_fragments: NibbleVec) -> Self::Result

Run the traversal, returning the result.

Let key_fragments be the bits of the key which are valid for insertion below the current node such that the 0th element of key_fragments describes the bucket that this key would be inserted into.