Crate radix_trie

A wonderful, fast, safe, generic radix trie implementation.

To get started, see the docs for Trie below.

Structs

Iter
Iterator over the keys and values of a Trie.

Keys
Iterator over the keys of a Trie.

NibbleVec
A data-structure for storing a sequence of 4-bit values.

Trie
Data-structure for storing and querying string-like keys and associated values.

Values
Iterator over the values of a Trie.

Traits

Traversal
Trait capturing a (mutable) traversal of a trie.

TrieKey
Trait for types which can be used to key a Radix Trie.