Struct radix_trie::NibbleVec

pub struct NibbleVec<A> where
    A: Array<Item = u8>,  { /* fields omitted */ }

A data-structure for storing a sequence of 4-bit values.

Values are stored in a Vec<u8>, with two values per byte.

Values at even indices are stored in the most-significant half of their byte, while values at odd indices are stored in the least-significant half.

Imagine a vector of MSB first bytes, and you’ll be right.

n = [_ _ | _ _ | _ _]

Implementations

impl<A> NibbleVec<A> where
A: Array<Item = u8>,

pub fn new() -> NibbleVec<A>

Create an empty nibble vector.

pub fn from_byte_vec(vec: Vec<u8, Global>) -> NibbleVec<A>

Create a nibble vector from a vector of bytes.

Each byte is split into two 4-bit entries (MSB, LSB).

pub fn as_bytes(&self) -> &[u8]

Returns a byte slice of the nibble vector’s contents.

pub fn into_bytes(self) -> Vec<u8, Global>

Converts a nibble vector into a byte vector.

This consumes the nibble vector, so we do not need to copy its contents.

pub fn len(&self) -> usize

Get the number of elements stored in the vector.

pub fn is_empty(&self) -> bool

Returns true if the nibble vector has a length of 0.

pub fn get(&self, idx: usize) -> u8

Fetch a single entry from the vector.

Guaranteed to be a value in the interval [0, 15].

Panics if idx >= self.len().

pub fn push(&mut self, val: u8)

Add a single nibble to the vector.

Only the 4 least-significant bits of the value are used.

pub fn split(&mut self, idx: usize) -> NibbleVec<A>

Split the vector into two parts.

All elements at or following the given index are returned in a new NibbleVec, with exactly idx elements remaining in this vector.

Panics if idx > self.len().

pub fn join(self, other: &NibbleVec<A>) -> NibbleVec<A>

Append another nibble vector whilst consuming this vector.