Struct radix_trie::Trie

pub struct Trie<K, V> {
    // some fields omitted
}

Tries allow collections of string-like keys to be efficiently stored and queried.

Any keys which share a common prefix are stored below a single copy of that prefix. This saves space, and also allows the longest prefix of any given key to be found.

You can read more about Radix Tries on Wikipedia.

Methods

impl<K, V> Trie<K, V> where K: TrieKey

fn new() -> Trie<K, V>

Create an empty Trie with no data.

fn len(&self) -> usize

Fetch the number of key-value pairs stored in the Trie.

fn is_empty(&self) -> bool

Determine if the Trie contains 0 key-value pairs.

fn get(&self, key: &K) -> Option<&V>

Fetch a reference to the given key's corresponding value (if any).

fn get_mut(&mut self, key: &K) -> Option<&mut V>

Fetch a mutable reference to the given key's corresponding value (if any).

fn insert(&mut self, key: K, value: V) -> Option<V>

Insert the given key-value pair, returning any previous value associated with the key.

fn remove(&mut self, key: &K) -> Option<V>

Remove and return the value associated with the given key.